rough golem death working
okay! this is my last push tonight for the golem portion of #25, which also
addresses #24 a bunch. by moving golems into objects held in game memory, file
i/o is drastically cut down (i'm sure there were like half a dozen file reads
for every golem action in the past); downside, of course, is the risk of losing
game data if there's an unclean exit (see #27, just opened as i was typing
this).

there's still a ton of bugs but it's 2am and i'm sleepy and can't think straight
anymore! but, golem expiration seems to correctly remove a golem from the game
objects and also delete the golem's save file.

major problems are mostly with legacy support for unrefactored code, and also
because i'm sleepy.

untested:
    * whether acquiring the golem drops works

probably broken:
    * checking if a player can afford a golem

still to implement:
    * golem autosave (really any autosave)
    * golem striking
    * golem decay

(now that i'm getting a clearer picture of this new engine, i think i have a
better understanding of making a global game state tick)